+++ library/src/methods/safeParse/types.ts
@@ -11,15 +11,41 @@ import type {
  */
 export type SafeParseResult<TSchema extends BaseSchema | BaseSchemaAsync> =
   | {
+      typed: true;
       success: true;
       /**
        * @deprecated Please use `.output` instead.
        */
       data: Output<TSchema>;
       output: Output<TSchema>;
+      /**
+       * @deprecated Please use `.issues` instead.
+       */
+      error: undefined;
+      issues: undefined;
     }
   | {
+      typed: true;
       success: false;
+      /**
+       * @deprecated Please use `.output` instead.
+       */
+      data: Output<TSchema>;
+      output: Output<TSchema>;
+      /**
+       * @deprecated Please use `.issues` instead.
+       */
+      error: ValiError;
+      issues: Issues;
+    }
+  | {
+      typed: false;
+      success: false;
+      /**
+       * @deprecated Please use `.output` instead.
+       */
+      data: unknown;
+      output: unknown;
       /**
        * @deprecated Please use `.issues` instead.
        */