+++ library/src/schemas/object/object.test.ts
@@ -2,6 +2,7 @@ import { describe, expect, test } from 'vitest';
 import { type ValiError } from '../../error/index.ts';
 import { parse } from '../../methods/index.ts';
 import { toCustom } from '../../transformations/index.ts';
+import { custom, minLength } from '../../validations/index.ts';
 import { never } from '../never/index.ts';
 import { number } from '../number/index.ts';
 import { optional } from '../optional/index.ts';
@@ -161,5 +162,73 @@ describe('object', () => {
     );
     expect(output1).toEqual(transformInput());
     expect(output2).toEqual(transformInput());
+  });
+
+  test('should execute pipe if output is typed', () => {
+    const requirement = (value: { key: string }) => value.key.length >= 10;
+    const schema = object({ key: string([minLength(10)]) }, [
+      custom(requirement),
+    ]);
+    const input = { key: '12345' };
+    const result = schema._parse(input);
+    expect(result).toEqual({
+      typed: true,
+      output: input,
+      issues: [
+        {
+          reason: 'string',
+          validation: 'min_length',
+          origin: 'value',
+          message: 'Invalid length',
+          input: input.key,
+          requirement: 10,
+          path: [
+            {
+              type: 'object',
+              input: input,
+              key: 'key',
+              value: input.key,
+            },
+          ],
+        },
+        {
+          reason: 'object',
+          validation: 'custom',
+          origin: 'value',
+          message: 'Invalid input',
+          input: input,
+          requirement,
+        },
+      ],
+    });
+  });
+
+  test('should skip pipe if output is not typed', () => {
+    const schema = object({ key: string() }, [
+      custom((value) => value.key.length >= 10),
+    ]);
+    const input = { key: 12345 };
+    const result = schema._parse(input);
+    expect(result).toEqual({
+      typed: false,
+      output: input,
+      issues: [
+        {
+          reason: 'type',
+          validation: 'string',
+          origin: 'value',
+          message: 'Invalid type',
+          input: input.key,
+          path: [
+            {
+              type: 'object',
+              input: input,
+              key: 'key',
+              value: input.key,
+            },
+          ],
+        },
+      ],
+    });
   });
 });