+++ library/src/utils/errorMessage/errorMessage.ts
@@ -1,0 +1,12 @@
+import type { ErrorMessage } from '../../types/index.ts';
+
+/**
+ * Returns the final string of the error message.
+ *
+ * @param message The error message.
+ *
+ * @returns The error message.
+ */
+export function errorMessage(message: ErrorMessage): string {
+  return typeof message === 'function' ? message() : message;
+}