+++ library/src/validations/isoTimestamp/isoTimestamp.ts
@@ -1,6 +1,6 @@
 import { ISO_TIMESTAMP_REGEX } from '../../regex.ts';
 import type { BaseValidation, ErrorMessage } from '../../types/index.ts';
-import { getOutput, getPipeIssues } from '../../utils/index.ts';
+import { actionIssue, actionOutput } from '../../utils/index.ts';
 
 /**
  * ISO timestamp validation type.
@@ -40,8 +40,8 @@ export function isoTimestamp<TInput extends string>(
     requirement: ISO_TIMESTAMP_REGEX,
     _parse(input) {
       return !this.requirement.test(input)
-        ? getPipeIssues(this.type, this.message, input, this.requirement)
-        : getOutput(input);
+        ? actionIssue(this.type, this.message, input, this.requirement)
+        : actionOutput(input);
     },
   };
 }